Refactor permissions' checks order
